Updated chirp controller to use and validate the actual user

+++ chirper-backend/src/controller/chirp-controller.ts
@@ -1,3 +1,4 @@
+import User from "@/model/user";
 import { createChirp, deleteChirpById, getAllChirps, getChirpById, updateChirpById } from "@/service/chirp-service";
 import express from "express";
 import { Request, Response } from "express";
@@ -30,8 +31,12 @@ router.get("/:id", async (req: Request, res: Response) => {
 router.post("/", async (req: Request, res: Response) => {
     const { text_content } = req.body;
 
-    // TODO: Dummy user id, need to implement authentication
-    const uid = 1;
+    const uid = (req.user as User).id;
+    if (uid === undefined) {
+        // User is not authenticated, so do not action the request
+        res.status(401).json({ error: "Unauthorized" });
+        return;
+    }
     try {
         const newChirp = await createChirp({ text_content, user_id: uid });
         res.status(201).json(newChirp);
@@ -43,6 +48,21 @@ router.post("/", async (req: Request, res: Response) => {
 router.put("/:id", async (req: Request, res: Response) => {
     const { id } = req.params;
     const { text_content } = req.body;
+
+    const uid = (req.user as User).id;
+
+    if (uid === undefined) {
+        // User is not authenticated, so do not action the request
+        res.status(401).json({ error: "Unauthorized" });
+        return;
+    }
+
+    if (uid !== (await getChirpById(parseInt(id))).user_id) {
+        // User is not the owner of the chirp, so do not action the request
+        res.status(403).json({ error: "Forbidden" });
+        return;
+    }
+
     try {
         const updatedChirp = await updateChirpById(parseInt(id), { text_content });
         if (updatedChirp) {
@@ -57,6 +77,21 @@ router.put("/:id", async (req: Request, res: Response) => {
 
 router.delete("/:id", async (req: Request, res: Response) => {
     const { id } = req.params;
+
+    const uid = (req.user as User).id;
+
+    if (uid === undefined) {
+        // User is not authenticated, so do not action the request
+        res.status(401).json({ error: "Unauthorized" });
+        return;
+    }
+
+    if (uid !== (await getChirpById(parseInt(id))).user_id) {
+        // User is not the owner of the chirp, so do not action the request
+        res.status(403).json({ error: "Forbidden" });
+        return;
+    }
+    
     try {
         const deletedChirp = await deleteChirpById(parseInt(id));
         if (deletedChirp) {
